aws-dbt2-aurora: save system logs from driver system
Adding the get-system-details role to capture system logs requires a
little juggling of how the results_directory was used.  So to use
results_directory in a more generic way, it seems like using synchronize
to retrieve the dbt2 results let's us do that as well as reduce the
number of tasks in the playbook by removing unnecessary archiving of
data.